fact-checker/requirements.txt
# Fact-Checker Dependencies

# Core AI/ML Libraries
langgraph>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0

# Web Scraping and Crawling
crawl4ai>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Text Processing
nltk>=3.8.0
spacy>=3.7.0
textblob>=0.17.0

# Utilities
pydantic>=2.0.0
httpx>=0.25.0
tenacity>=8.2.0
tqdm>=4.66.0

# JSON and YAML Support
pyyaml>=6.0.0
jsonschema>=4.19.0

# Logging and Monitoring
structlog>=23.1.0
rich>=13.5.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
